Change getTransactionHash to use a unique hash vs `this`
For state channels, there are situations where you need to use a counterfactual multisignature wallet and so you won't know the address ahead of time. You will know the counterfactual address, however, and so you can use it when calling `setup` to allow for transactions that are signed to be valid when the contract is put on chain. This forces the user to handle replay-protection though, and not the blockchain implicitly.