Update ModuleManager.sol
Based on the setupModules and enableModule methods in the ModuleManager, we can ascertain that SENTINEL_MODULES points to the head node. The linked list can be represented as follows: SENTINEL_MODULES --> newModule --> ModuleA --> ModuleB --> ModuleC --> address(0x01).

Given this, isn't there redundancy in the getModulesPaginated method? For example, the check next != SENTINEL_MODULES seems redundant because next can never be SENTINEL_MODULES as it points to the head node. Moreover, the first line of the getModulesPaginated method already ensures that start != SENTINEL_MODULES.