Audit Changes for `v1.4.1-2` cherry picked from `release/v1.4.1-2` (#815)
This PR makes the changes based on Certora's audit report from the
branch `release/v1.4.1-2` to `main`.

In the release branch, we used the `ISafe` contract written inside the
respective library files, but with `main`, we use the `ISafe` standard
interface in the `interfaces` folder. Similarly, `Enum` was moved from
`common` to `libraries`, and thus that is used for the same in the
migration contracts.

Note: A `fmt` change was made to `DebugTransactionGuard.sol`.