Release 1.4.1-3 (#841)
This pull request includes updates to the `CHANGELOG.md` and
`package.json` files to document the latest changes and support for
zkSync. The most important changes include adding new contracts for
migration and zkSync support and updates to the version number.

### Documentation Updates:

*
[`CHANGELOG.md`](diffhunk://#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4edR7-R11):
Added release iterations for versions 1.4.1-2 and 1.4.1-3, including new
contracts for Safe migrations and zkSync support.
*
[`CHANGELOG.md`](diffhunk://#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4edL17-R46):
Updated the addresses for core, factory, handler, lib, and storage
reader contracts to include both EVM and zkSync addresses.

### Version Update:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L3-R3):
Updated the version number from 1.4.1-2 to 1.4.1-3.