[Certora Audit] G-05. Use `iszero` instead of `eq(*, 0)` (#892)
This pull request includes a critical update to the `SafeProxyFactory`
contract in the `contracts/proxies/SafeProxyFactory.sol` file. The
change improves the gas cost of the assembly code by replacing the `eq`
function with the `iszero` function.

Key change:

*
[`contracts/proxies/SafeProxyFactory.sol`](diffhunk://#diff-53aac98a01e16743466b00bdcb233208f95e1113cc90b95181187d9862b4ad6eL43-R43):
Replaced the `eq` function with the `iszero` function in the assembly
block to improve code readability and reliability.