`v1.5.0` Audit (#912)
This pull request includes documentation updates related to the audit
results for version 1.5.0. The changes provide details about the audit
process, including the auditor, notes on the audit, and the final audit
report file.

Documentation updates:

*
[`docs/audit_1_5_0.md`](diffhunk://#diff-16a7a881e174c77f794e98d3b228ce159cbc145d6541b331ef2f90bd3c2616bcR1-R15):
Added a new section titled "Audit Results" which includes information
about the auditor (Certora), notes on the audit, and a link to the final
audit report for version 1.5.0.