Separate zkSync Tests to Separate CI Step
This PR moves the zkSync tests outside of the normal `test` NPM
script. Since this would make it no longer run by default in CI, we
added a separate `tests-zk` step to run the zkSync tests only.

This split allows us to potentially merge PRs when zkSync tests fail
as long as the normal test suite does not. It is not the first time
that CI starts failing without cause because of zkSync tests. As far
as I can tell, the zkSync tooling downloads an in-memory node binary
that it starts for actually executing the tests. In the past, tests
have randomly started to fail because of:

* Changes to the behaviour of the in-memory node ([example](https://github.com/safe-global/safe-smart-account/actions/runs/14359540825/job/40257156305?pr=938))
* Binary no longer hosted on the same URL ([example](https://github.com/safe-global/safe-smart-account/actions/runs/14359540825/job/40257156305?pr=938))

When this happens, we do not want to block changes that are unrelated
to contract logic (as they should not, in theory, have any affect on
the zkSync tests).

Furthermore, running the zkSync tests takes quite a bit longer than the
regular tests, so this also makes local `npm test` a bit faster.