Ackee[W6]: Document `TokenCallbackHandler` Locks Tokens
A byproduct of the `TokenCallbackHandler` implementing token callback
functions is that it is able to receive tokens (such as ERC-721)
despite not being designed to do so. Tokens sent to this contract will
be permanently locked: SO DO NOT SEND TOKENS TO THIS CONTRACT!

This PR adds a documentation blurb to the contract in question to
document this shortcomming. Unfortunately, this is not something that
can be reasonably checked for such that the handlers `revert`.