Ackee[I2]: Make `ISafe` Payable (#967)
`ISafe` interface was not payable, but it should be as the `Safe` can
receive tokens. This PR adds a `INativeCurrencyPaymentFallback`
interface with a payable `receive` function and makes the `ISafe`
interface inherit from it. This makes the `ISafe` interface payable, and
requires addresses that are cast to it to also be `payable`.