Bump Hardhat Toolbox Version
This is a follow up to #1000, where we finally bump the hardhat
toolbox verion to v6. Notable the _only_ change that that entails is
going from `hardhat-gas-reporter` `~1` to `~2`, which enables the gas
reporting by default.

After some more digging, I found that simply disbling the gas
reporting would cause tests to fail (with errors: "calling done()
twice on test" and "cannot index property to on undefined"). I
narrowed the issue down to a specific test, which is now disabled when
gas reporting is on. Furthermore, I disabled gas reporting by default,
as I find it quite noisy. To run tests with a gas report:

```sh
HARDHAT_ENABLE_GAS_REPORTER=1 npm test
```