Bump Dependency Versions (#1000)
Just update dependencies to their latest and `npm update`.

Note that we *DO NOT* upgrade `@nomicfoundation/hardhat-toolbox` to
v6.0.0 because it is causing weird intermittent test failures ("calling
`done()` twice on test" and "cannot index property `to` on
`undefined`"). I tried debugging why exactly this is happening, but I
didn't figure it out (the stack trace is not helpful, so I have no idea
_where_ it is happening).

I did, however, narrow it down to a few tests: `SafeToL2Setup` +
`SafeToL2Migration` tests enabled **together** trigger the error (but
not when running either suite in isolation). I will debug a bit further,
but for now here are the dependency bumps for other packages.