Deprecate `SafeToL2Migration`
The `SafeToL2Migration` contract was previously included in the Safe
v1.5.0 release, but it had not been modified to work with the Safe
v1.5.0 contracts. In particular, the version string "1.5.0" had not
been allow-listed by the contract for migrations.

The `SafeToL2Migration` contract is written in such a way that it
requires that the version of the L1 singleton matches the L2 singleton
for the migration to succeed. As such, adding support for the 1.5.0
singletons would have arguably very limited benefit, as all recent
Safes are created with the `SafeToL2Setup` contract in order to ensure
consistent addresses across Ethereum Mainnet and L2s (which use a
different singleton). As such, there should be no new Safes created by
the Safe{Wallet} interface that would require migration, and
therefore, there would be no v1.5.0 Safes created by the interface
that would require migration, making the contract no longer useful for
future versions. **Therefore, we have chosen to deprecate the
`SafeToL2Migration` contract for the 1.5.0 release**.

There is the possibility of adding migration contract from Safes with
singleton version `X.Y.Z+L1` to `1.5.0+L2`, but without clear
requirements, it does not have to be added now or affect the release
of the 1.5.0 contracts.