Docs: Update the `IModuleManager` and `IOwnerManager` interfaces to clarify the usage of the `prevModule` and `prevOwner`  (#993)
Updated the `IModuleManager` and `IOwnerManager` interfaces to clarify
the usage of the `prevModule` and `prevOwner` parameters in the
`disableModule` and `removeOwner` functions. Added details specifying
that when the module or owner to be removed is the first (or only)
element in the list, the respective `prevModule` or `prevOwner` must be
set to the sentinel address `0x1`. This improves understanding of the
linked list structure used in these functions.

fixes https://github.com/safe-global/safe-smart-account/issues/992

---------

Co-authored-by: Nicholas Rodrigues Lordello <n@lordello.net>
Co-authored-by: Shebin John <admin@remedcu.com>