Creation Code Hash (#1019)
# TLDR
- Adds the creation code hash in Safe Proxy Factory.

## LLM Description
This pull request introduces a new method in the `SafeProxyFactory`
contract to enhance functionality by providing the ability to retrieve
the creation code hash of a `SafeProxy`. This addition can be used to
compute a `SafeProxy` creation address.

Enhancements to `SafeProxyFactory`:

*
[`contracts/proxies/SafeProxyFactory.sol`](diffhunk://#diff-53aac98a01e16743466b00bdcb233208f95e1113cc90b95181187d9862b4ad6eR38-R45):
Added the `proxyCreationCodehash` method, which returns the keccak256
hash of the `SafeProxy` creation code. This method is marked as `public`
and `pure`, and it can be used to compute the creation address of a
`SafeProxy`.

Fixes #994